requirements.txt
amqp==1.4.6
anyjson==0.3.3
asgiref==3.3.4
billiard==3.3.0.20
celery==3.1.18
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
kombu==3.0.30
pkg-resources==0.0.0
pytz==2015.4
redis==2.10.3
requests==2.7.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
